Introduce Commands for creating calls
Commands are a good way of providing an expectation to modify or add
something to the database where you may not have all the fields.

For example, the CreateHelpRequestCall command only contains the
information required to create a HelpRequest Call. It does not contain
things like the id, because this cannot be known before it has been
persisted.

You can see an example of why this is important in the E2E test.
Previously, if an API consumer decided to send through a call with a
speific id, it'd be stored in the database with that id - this is
dangerous.