Introduce publish-to-swagger-hub script
Given a stage it will generate then publish the documentation into
swaggerhub. In an ideal world that will be driven via docker - but
that's something we can come back to at a later date.

Next step is to integrate this into the deployment script so that
Swaggerhub is a clear reflection of the currently deployed APIs