Fixed minor oversights in FindResident tests
New day gives a fresh pair of eyes to look back at the code - found couple errors (due to copy & paste) that missed my eyes the other day.

Mostly like tests are named, as well as couple lines of setup that had set up a matching fields, they shouldn't have done so.. and a missing variable in the test cases.