Fixed the FindResident Dob not empty check & added more extensive not empty check
Previously it would evaluate Fname, Lname and Dob when:

At least 1 of 3 parts of Dob was present (or operator).

Changed to require all 3 parts of Dob as I think that knowing Fname, Lname and DobDay is not enough to guarantee it's a duplicate resident.

More extensive not empty check as dates sometimes come in as empty strings