Fixed FindResident behaviour on empty First and Last names.
Made it so it wouldn't do any further comparisons when the First name or last name is not present.

It's because I don't think having just the:
1) Uprn & (FirstName | LastName)
2) Dob & (FirstName | LastName)

allows for giving a guarantee that a person is a duplicate. What was happening before, essentially the whitespaces would manage to pass through to the DB due to lack of validation, and if we have record with LastName = "" and and insert command with LastName = "", then these 2 match, contributing to the potentially false identification of a resident as a duplicate.

If within if so that I would only need to check the names being empty once.