FindResident - Added a deduplication check by NhsNumber
Every resident gets a unique NhsNumber. They can have more than 1 NhsNumber, however it's guaranteed that no other resident has the same NhsNumber. (and the collision chance is very low)

Due to this checking just the NhsNumber match is enough to guarantee that it's a duplicate resident. Hence it goes as a first rule before any other checks.

Note! NhsNumber not matching does not mean resident is not a duplicate. It's because, like mentioned before, any resident can have more than 1 such number assigned to them. So whatever is in the create resident command, might just be another unknown to us NhsNumber of potentially already existing in DB resident. Hence, we do further checks, to identify whether one from command is unique or not.

Made this check outside the checking for names being empty block as, like mentioned before, matching NhsNumber guarantees it's duplicate resident, so no need to require for names as any info in here will just get appended.

(Question however... what if the record in DB does not have the first name, but the data coming in through create request has matching NhsNumber & the missing First Name? Does our creation part of the flow populate the missing name?)