Upgraded the Phone number de-duplication rule to check both number fields.
Did this as sometimes only mobile number is provided, but for some reason it's the wrong field (like desk phone).

It's not an issue doing such check as landline and mobile phone numbers do not clash, so this matching rule upgrade cannot produce false positives that way.

Of course, if the same thing was done with DobMonth.& DobDay, that would be a different case. As due to the way month is stored, it overlaps with DobDay values. Example: April is stored as 04, so it would overlap with the 4th day of month (04), hence possibly creating a false positive This is not the case with phone numbers.